refactor: make tests compare formatted AST
make the tests for an ast rely on an external method that iterates over
the tree and prints it in a readable manner instead of relying on :==
logic that is currently untested in isolation.

The tests before were effectively using the code they were verifying to
verify the code in a cyclic pattern.