feat: add division
IDIV command uses RDX register so using it as the divisor leads to a
floating point error as it is zeroed presumably before the division
leading to a division by 0. for this reason the RDX register has been
changed to the RDI register. This might need to be changed soon.