refactor: generic return register and param order
all registers used have been canged around so that all method in general
return their result in the RAX register. Everything going mainly through
the RAX register is to simplify nested structures.

The order of parameters is done with the idea that it hopefullylead to
easier extendability in the future. This will mean that side effects
might be counter intuitive however the main goal is for a lack of side
effects.