refactor: make lex tests more readable on failure
changed the token comparison to compare two string generated from the
expected and the actual arrays of tokens. This means that if they don't
match, the differenc is more clearly highlighted instead of rspec just
saying that there is an error somewhere in the array which isn't very
helpful when trying to assess where things have gone wrong.